+++ packages/core/src/components/accordion/components/accordion-section.interfaces.ts
@@ -2,6 +2,8 @@ export type AccordionSectionState = "success" | "info" | "warning" | "danger" | 
 
 export type AccordionHeading = "h2" | "h3" | "h4" | "h5";
 
+export type AccordionSectionWijzigactie = "voegtoe" | "verwijder";
+
 export const stateMap: Record<AccordionSectionState, string> = {
   success: "succes:",
   info: "info:",